Todos los sliders de rango terminados a falta de estilos

+++ src/app/informe-filters/components/gradient-filter/gradient-filter.component.ts
@@ -3,8 +3,11 @@ import { Component, OnInit } from '@angular/core';
 import { MatSliderChange } from '@angular/material/slider';
 
 import { FilterService } from '@core/services/filter.service';
+import { PcService } from '@core/services/pc.service';
 
 import { GradientFilter } from '@core/models/gradientFilter';
+
+import { Options } from '@angular-slider/ngx-slider';
 
 @Component({
   selector: 'app-gradient-filter',
@@ -15,26 +18,23 @@ export class GradientFilterComponent implements OnInit {
   minGradiente: number;
   maxGradiente: number;
   rangoMinGradiente: number;
+  rangoMaxGradiente: number;
   filtroGradiente: GradientFilter;
+  options: Options;
 
-  constructor(private filterService: FilterService) {}
+  constructor(private filterService: FilterService, private pcService: PcService) {
+    this.minGradiente = this.pcService.getMinGradienteNormalizado();
+    this.maxGradiente = this.pcService.getMaxGradienteNormalizado();
+    this.rangoMinGradiente = this.minGradiente;
+    this.rangoMaxGradiente = this.maxGradiente;
+    this.options = { floor: this.minGradiente, ceil: this.maxGradiente };
+  }
 
   ngOnInit(): void {}
 
-  formatLabel(value: number | null) {
-    if (!value) {
-      return this.rangoMinGradiente;
-    }
-    return value + 'ºC';
-  }
-
-  onInputFiltroGradiente(event: MatSliderChange) {
-    this.rangoMinGradiente = event.value;
-  }
-
   onChangeFiltroGradiente() {
-    this.filtroGradiente = new GradientFilter('gradient', this.rangoMinGradiente, this.maxGradiente);
-    if (this.rangoMinGradiente === this.minGradiente) {
+    this.filtroGradiente = new GradientFilter('gradient', this.rangoMinGradiente, this.rangoMaxGradiente);
+    if (this.rangoMinGradiente === this.minGradiente && this.rangoMaxGradiente === this.maxGradiente) {
       // si se selecciona el mínimo desactivamos el filtro ...
       this.filterService.deleteFilter(this.filtroGradiente);
     } else {
